[frontend] edit everything as QA sent

+++ backend/DB/migrations/20201127120719_delete_imgurl_from_student.js
@@ -1,7 +1,8 @@
 
 exports.up = function(knex) {
     return knex.schema.table('student', function(table) {
-        table.dropColumn('imgurl')
+      if(knex.schema.hasColumn(table, "imgurl")) {table.dropColumn('imgurl') }  
+      
       })
 };
